project/requirements.txt
Django==2.0.2
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djoser==1.1.5
pkg-resources==0.0.0
PyJWT==1.5.3
pytz==2017.3
